+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed Txn and Query{} API, reduced api noise, removed funcs that should not be used now after upserts and types got introduced
 * Added SetRDF and DeleteRDF mutation helpers, along with Query{} helpers
 * Renamed `FlattenJSON` -> `FlattenRespToObject` to better describe what it does
+* Renamed `QueryJSON` -> `QueryDQL`
 ---
 
 ## v3.1.0 - 2020-11-23